requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Login==0.6.3
gcloud==0.18.3
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauth2client==4.1.3
protobuf==5.28.0rc1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==72.1.0
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==3.0.3
WTForms==3.1.2
